% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bids}
\alias{Bids}
\title{Bids Received by U.S. Firms}
\format{
A data frame with 126 rows and 13 variables: \describe{
  \item{docno}{doc no.} \item{weeks}{weeks} \item{numbids}{count}
  \item{takeover}{delta(1 if taken over)} \item{bidprem}{bid Premium}
  \item{insthold}{institutional holdings} \item{size}{size measured in
  billions} \item{leglrest}{legal restructuring} \item{rearest}{real
  restructuring} \item{finrest}{financial restructuring}
  \item{regulatn}{regulation} \item{whtknght}{white knight} }
}
\source{
Jaggia, Sanjiv and Satish Thosar (1993) “Multiple Bids as a
  Consequence of Target Management Resistance”, Review of Quantitative
  Finance and Accounting, 447–457.

  Cameron, A.C. and Per Johansson (1997) “Count Data Regression Models using
  Series Expansions: with Applications”, Journal of Applied Econometrics, 12,
  may, 203–223.
}
\usage{
Bids
}
\description{
A dataset with bids received by U.S. firms.
}
\keyword{datasets}
