% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CustomerProfile}
\alias{CustomerProfile}
\title{Customer profile for a household supplies company}
\format{
A data frame with 110 rows and 6 variables: \describe{
  \item{ncust}{number of customer of the census tracts who visit the store.}
  \item{nhu}{number of housing units in the census tracts} \item{aid}{average
  income in dollars} \item{aha}{average housing unit in years}
  \item{dnc}{distance to the nearest competitor in miles} \item{ds}{distance
  to store in miles} }
}
\source{
http://www.leg.ufpr.br/lib/exe/fetch.php/publications:papercompanions:ptwdataset4.txt
}
\usage{
CustomerProfile
}
\description{
An observation corresponds to the census tracts within 10-mile radius around
certain store.
}
\keyword{datasets}
