% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGLMExtPois-package.R
\docType{package}
\name{DGLMExtPois-package}
\alias{DGLMExtPois}
\alias{DGLMExtPois-package}
\title{DGLMExtPois: Double Generalized Linear Models Extending Poisson Regression}
\description{
Model estimation, dispersion testing and diagnosis of hyper-Poisson Saez-Castillo, A.J. and Conde-Sanchez, A. (2013) \doi{10.1016/j.csda.2012.12.009} and Conway-Maxwell-Poisson Huang, A. (2017) regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/franciscomartinezdelrio/DGLMExtPois}
  \item Report bugs at \url{https://github.com/franciscomartinezdelrio/DGLMExtPois/issues}
}

}
\author{
\strong{Maintainer}: Francisco Martinez \email{fmartin@ujaen.es}

Authors:
\itemize{
  \item Antonio Jose Saez-Castillo \email{ajsaez@ujaen.es}
  \item Antonio Conde-Sanchez \email{aconde@ujaen.es}
}

}
\keyword{internal}
