% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGLMExtPois.R
\docType{package}
\name{DGLMExtPois}
\alias{DGLMExtPois}
\title{DGLMExtPois: Double Generalized Linear Models Extending Poisson Regression.}
\description{
The DGLMExtPois package provides two categories of important functions:
glm.hP and glm.CMP.
}
